Fix for 7200/7500 hardware mixup
For some reason the system thinks the grass tower is running a 7500 not a 7200. This removes the flowdrive variable from the summary files and makes a large chunk of this code obsolete. While we work out a fix to the problem this will allow us to keep an eye on the tower.